Add code for pipelines and activations.
