+++ src/rlhf_model_training/model_pipelines/hh_training_pipeline.py
@@ -1,0 +1,6 @@
+from src.rlhf_model_training.model_pipelines.rlhf_model_pipeline import RLHFModelPipeline
+
+class HH_TrainingPipeline(RLHFModelPipeline):
+    """
+    Extends DPOModelPipeline
+    """
