+++ src/rlhf_model_training/model_pipelines/unalignment_training_pipeline.py
@@ -1,0 +1,6 @@
+from src.rlhf_model_training.model_pipelines.rlhf_model_pipeline import DPOModelPipeline
+
+class Unalignment_TrainingPipeline(DPOModelPipeline):
+    """
+    Extends DPOModelPipeline
+    """
